Added PointerStack.units member.
Could have sworn we had something like this in there. Just goes to show, watch what you say! I give a fuck! (Isn't that quite odd to hear in the inverse magnitude with attitude? That's because, nobody gives a fuck! :)

This member value is used to store how many total units are available to the stack without reallocation.

Modified: src/pointer-stack.c