Added terminal coloring.
Gedit terminal plugin sucks... You may not be able to see the color output because the software is not written correctly. Might add a --gedit option but its not likely to support foolishness today, tomorrow or next week.

Use a real terminal if it matters to you.

Modified: src/test/suite.sh